Add basic loop implementation
