+++ tests/WorkflowTestTrait.php
@@ -6,6 +6,7 @@ namespace PHPWorkflow\Tests;
 
 use PHPWorkflow\State\WorkflowContainer;
 use PHPWorkflow\State\WorkflowResult;
+use PHPWorkflow\Step\LoopControl;
 use PHPWorkflow\Step\WorkflowStep;
 use PHPWorkflow\WorkflowControl;
 
@@ -40,16 +41,37 @@ trait WorkflowTestTrait
         };
     }
 
+    private function setupLoop(string $description, callable $callable): LoopControl
+    {
+        return new class ($description, $callable) implements LoopControl {
+            private string $description;
+            private $callable;
+
+            public function __construct(string $description, callable $callable)
+            {
+                $this->description = $description;
+                $this->callable = $callable;
+            }
+
+            public function getDescription(): string
+            {
+                return $this->description;
+            }
+            public function executeNextIteration(
+                int $iteration,
+                WorkflowControl $control,
+                WorkflowContainer $container
+            ): bool {
+                return ($this->callable)($control, $container);
+            }
+        };
+    }
+
     private function assertDebugLog(string $expected, WorkflowResult $result): void
     {
-        $this->assertSame($expected, preg_replace('#[\w\\\\]+@anonymous[^)]+#', 'anonClass', preg_replace('/[\d.]+ms/', '*', $result->debug())));
-
-        return;
-
-        $this->assertSame($expected, preg_replace(
-            ['/[\d.]+ms/', '/[\w\\]+@anonymous[\w\\/:\-\.\$]+/'],
-            ['*',          'anonClass'],
-            $result->debug(),
-        ));
+        $this->assertSame(
+            $expected,
+            preg_replace('#[\w\\\\]+@anonymous[^)]+#', 'anonClass', preg_replace('/[\d.]+ms/', '*', $result->debug())),
+        );
     }
 }
